Extract out slider
